Move out shared types
