+++ packages/web/src/pages/upload-page/fields/RemixSettingsField/RemixSettingsField.tsx
@@ -21,7 +21,7 @@ import { fullTrackPage } from 'utils/route'
 
 import { useTrackField } from '../../hooks'
 import { SingleTrackEditValues } from '../../types'
-import { IS_STREAM_GATED, STREAM_CONDITIONS } from '../AccessAndSaleField'
+import { IS_STREAM_GATED, STREAM_CONDITIONS } from '../types'
 
 import styles from './RemixSettingsField.module.css'
 import { RemixSettingsMenuFields } from './RemixSettingsMenuFields'
