+++ packages/web/src/pages/upload-page/fields/types.ts
@@ -1,6 +1,50 @@
+import {
+  AccessConditions,
+  FieldVisibility,
+  Nullable,
+  StreamTrackAvailabilityType
+} from '@audius/common'
+
 export type FieldProps = {
   name: string
   label: string
   required?: boolean
   errorMessage?: string
+}
+
+export enum SpecialAccessType {
+  TIP = 'tip',
+  FOLLOW = 'follow'
+}
+
+export const IS_SCHEDULED_RELEASE = 'is_scheduled_release'
+export const IS_UNLISTED = 'is_unlisted'
+export const IS_STREAM_GATED = 'is_stream_gated'
+export const STREAM_CONDITIONS = 'stream_conditions'
+export const IS_DOWNLOAD_GATED = 'is_download_gated'
+export const DOWNLOAD_CONDITIONS = 'download_conditions'
+export const STREAM_AVAILABILITY_TYPE = 'stream_availability_type'
+export const SPECIAL_ACCESS_TYPE = 'special_access_type'
+export const FIELD_VISIBILITY = 'field_visibility'
+export const PRICE = 'stream_conditions.usdc_purchase.price'
+export const PRICE_HUMANIZED = 'price_humanized'
+export const PREVIEW = 'preview_start_seconds'
+export const IS_DOWNLOADABLE = 'is_downloadable'
+export const IS_ORIGINAL_AVAILABLE = 'is_original_available'
+export const DOWNLOAD_REQUIRES_FOLLOW = 'download_requires_follow'
+export const STEMS = 'stems'
+export const DOWNLOAD = 'download'
+export const CID = 'download.cid'
+export const DOWNLOAD_AVAILABILITY_TYPE = 'download_availability_type'
+export const DOWNLOAD_PRICE = 'download_conditions.usdc_purchase.price'
+export const DOWNLOAD_PRICE_HUMANIZED = 'download_price_humanized'
+
+export type AccessAndSaleFormValues = {
+  [IS_UNLISTED]: boolean
+  [STREAM_AVAILABILITY_TYPE]: StreamTrackAvailabilityType
+  [STREAM_CONDITIONS]: Nullable<AccessConditions>
+  [SPECIAL_ACCESS_TYPE]: Nullable<SpecialAccessType>
+  [FIELD_VISIBILITY]: FieldVisibility
+  [PRICE_HUMANIZED]: string
+  [PREVIEW]?: number
 }
